move home and landing pages to check_login dir
