+++ lib/pages/check_login/landing.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:trail_buddies/widgets/common/button.dart';
-import '../widgets/common/background_image.dart';
-import '../widgets/landing_page/logo.dart';
+import '../../widgets/common/background_image.dart';
+import '../../widgets/landing_page/logo.dart';
 
 class LandingPage extends StatelessWidget {
   const LandingPage({Key? key}) : super(key: key);
